Remove unused non-cross-platform source files

+++ src/SexyAppFramework/misc/RegEmu.cpp
@@ -130,7 +130,7 @@ void regemu::SetRegFile(const std::string& fileName)
 	}
 
 	fclose(f);
-	printf("RegEmu: Loaded from '%s': %ld total key(s)\n", currFile.c_str(), registry.size());
+	printf("RegEmu: Loaded from '%s': %zu total key(s)\n", currFile.c_str(), static_cast<size_t>(registry.size()));
 }
 
 bool regemu::RegistryRead(const std::string& keyName, const std::string& valueName, uint32_t* type, uint8_t* value, uint32_t* length)
